+++ CHANGES.md
@@ -7,7 +7,15 @@
 * [**Donate** for the further development & support of the plugin](https://www.paypal.me/deckerweb)
 * [Plugin's documentation & FAQ](https://wordpress.org/plugins/builder-template-categories/#faq)
 
+
 ## Changelog of the Plugin
+
+### 0.9.1 - 2018-08-17
+
+* New: Added Code Snippets as GitHub Gists
+* New: Added short and easy "documentation" as Wiki in official GitHub repository for the plugin
+* New: Added integration "WPBakery Page Builder" (the old Visual Composer), via its "Templatera" Add-On plugin
+
 
 ### 0.9.0 - 2018-08-16